requirements.txt
streamlit==1.16.0
ultralytics==8.0.3
Pillow==9.3.0